% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPG_func.R
\name{insider}
\alias{insider}
\title{PCA and NN functions}
\usage{
insider(
  gi = gi,
  seu = seu,
  graph = graph,
  sp = sp,
  arr = arr,
  ct = ct,
  k = 10,
  cat = "group"
)
}
\value{
neighbourhood percentage connection matrix
}
\description{
Constructing PCA space and drawing NN graph using genes of interest
}
