% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPG_func.R
\name{lapply2}
\alias{lapply2}
\title{extension of lapply function}
\usage{
lapply2(X, FUN, ...)
}
\arguments{
\item{X}{vector names}

\item{FUN}{function in lapply}
}
\value{
A named vector
}
\description{
rewrite lapply function so that it returns a named vector
}
