% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPG_func.R
\name{nnfunc}
\alias{nnfunc}
\title{NN Connection}
\usage{
nnfunc(s, graph, label, k, cat)
}
\value{
The average score per cell type and average percentage per each cell
}
\description{
Calculating Percentage Neighbourhood Connection
}
