% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XPG_func.R
\name{XPG}
\alias{XPG}
\title{Main XPG function}
\usage{
XPG(g, sp, graph, seu, ct, core, k, cat)
}
\arguments{
\item{g}{Genes of interest}

\item{sp}{Starting panel gene list}

\item{graph}{Nearest Neighbour graph in Seurat object}

\item{seu}{Seurat object}

\item{ct}{celltype column in Seurat object}

\item{core}{number of available cores}

\item{k}{number neighbour per cell}

\item{cat}{average per cell type = 'group' or per cell = 'cell'}
}
\value{
a list of 3 objects (1. Percentage matrix 2. Ranked gene list 3. Time(s) per run)
}
\description{
Initialise all empty variables
}
\examples{
output = main_func(g = c('EPCAM','MS4A1','MZB1'), sp = c('KLF2','CD8A','COL1A1','EPCAM'), graph = 'RNA_nn', seu = obj, ct = 'celltype', core = 100 , k = 10, cat = 'group');
}
